Refactor command error types
We want to be able to tell if the errors are from a bad client input
or if they are server generated. The way to do this is to check the
parent class, instead of having properties that tell so.

This would allow infinite extensibility since different error types can
be parents of other errors. For example, for authorization errors.