Refactor the base handler to avoid useless boilerplate on functions
Having to specify two times the class of the input, output and handler
parameters of the function was boilerplate-y when we can retrieve
that information from the generic specialization of the base handler.

It's a runtime performance penalty to ease development. Tests should
be performed to check whether this is too much of a penalty for aws
lambda.